weather-arb-bot-12/requirements.txt
# HTTP
httpx>=0.27.0
tenacity>=8.2.3

# Polymarket
py-clob-client>=0.17.0

# Database
# sqlite3 is in stdlib, no install needed

# Scheduling
APScheduler>=3.10.4

# Configuration
python-dotenv>=1.0.1

# Dashboard
streamlit>=1.35.0
plotly>=5.22.0
pandas>=2.2.0

# Testing
pytest>=8.2.0
pytest-mock>=3.14.0
